# Adds index to improve dequeuing performance