@mookid8000 2020-06-12
Lovely! 👍 

It's out in Rebus.PostgreSql 7.1.1 on NuGet.org in a few minutes 🙂 

Thanks for contributing!